requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
apscheduler==3.10.4
python-dotenv==1.0.0
aiosqlite==0.19.0
asyncpg==0.29.0
aiomysql==0.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
